+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] - 2016-04-07
+### Added
+- these changes are thanks to [deckarep](https://github.com/deckarep). Thanks!
+- more updates to error naming convention
+- more error handing on HTTP request
+
 ## [1.0.1] - 2016-04-07
 ### Added
 - these changes are thanks to [deckarep](https://github.com/deckarep). Thanks!